Update logic for getting jobId.

+++ ernie-engine/src/main/scala/com/ksmpartners/ernie/engine/Messages.scala
@@ -11,9 +11,9 @@ import actors.Actor
 import com.ksmpartners.ernie.model.JobStatus
 
 case class ReportRequest(rtpDefId: String)
-case class JobRequest(rtpDefId: String, jobId: Int, self: Actor)
-case class StatusRequest(jobId: Int)
-case class Notify(jobId: Int, jobStatus: JobStatus, self: Actor)
+case class JobRequest(rtpDefId: String, jobId: Long, self: Actor)
+case class StatusRequest(jobId: Long)
+case class Notify(jobId: Long, jobStatus: JobStatus, self: Actor)
 
 case class ShutDownRequest()
 case class ShutDownResponse()
